Refactor reusable test classes into common module
